Remove Oak Functions Request Proto
I don't think we need/want this. The request proto is just a wrapper for the byte payload. It never actually reaches the Oak functions runtime, it's contents are read at the Client level, which somewhat counterintuitively just sends a the byte array directly.

Removing it means we can share the client between runtimes, without needing to also share all the invocation protos